+++ packages/zen-router-preact/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sylphlab/zen-router-preact
 
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/zen@3.4.0
+  - @sylphx/zen-router@3.0.5
+
 ## 3.0.4
 
 ### Patch Changes