+++ packages/zen-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sylph/router
 
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/zen@3.4.0
+  - @sylphx/zen-patterns@5.0.0
+
 ## 3.0.4
 
 ### Patch Changes